+++ R/table_continuous_lm_print.R
@@ -17,6 +17,7 @@ print.spicy_continuous_lm_table <- function(x, ...) {
   digits <- attr(x, "digits") %||% 2L
   fit_digits <- attr(x, "fit_digits") %||% 2L
   effect_size_digits <- attr(x, "effect_size_digits") %||% 2L
+  p_digits <- attr(x, "p_digits") %||% 3L
   decimal_mark <- attr(x, "decimal_mark") %||% "."
   ci_level <- attr(x, "ci_level") %||% 0.95
   by_label <- attr(x, "by_label") %||% "Predictor"
@@ -35,6 +36,7 @@ print.spicy_continuous_lm_table <- function(x, ...) {
     digits = digits,
     fit_digits = fit_digits,
     effect_size_digits = effect_size_digits,
+    p_digits = p_digits,
     decimal_mark = decimal_mark,
     ci_level = ci_level,
     show_statistic = show_statistic,